==> data_processing/network_assignment/code/10_12_Foreign_Matching_Foreign_Port_Add_Centroids.py <==
# =============================================================================
# this was done manually in QGIS.
# =============================================================================
# Adding New Nodes from an Existing Shapefile in QGIS
# Open QGIS and Load the Shapefile:

# Launch QGIS.
# Click on Layer > Add Layer > Add Vector Layer.
# Browse and select your shapefile. Click Open to load it into the map view.
# Start Editing the Layer:

# Right-click on the shapefile layer in the Layers panel and select Toggle Editing.
# The editing tools will become available.
# Add a New Node (Point):

# Select the Add Point Feature tool from the toolbar (this icon looks like a point).
# Click on the map where you want to add the new node.
# Fill in Attributes:

# After placing the point, a dialog box will pop up, prompting you to fill in the attribute data for the new node. Enter the required information and click OK.
# Save Edits:

# Once you’ve added the nodes, click the Save Layer Edits button (floppy disk icon) or right-click on the layer and choose Save Layer Edits.
# Stop Editing:

# Right-click on the shapefile layer and select Toggle Editing again to stop editing.
